+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 4.1.0 (2014-08-??)
 ------------------
 
-  * added `stubbles\lang\Parse`, unified string to value parsing for properties and annotation values
+  * added `stubbles\lang\Parse`
+    * unified string to value parsing for properties and annotation values
+    * properties bound via `stubbles\ioc\Binder::bindProperties()` are now injected as parsed values instead of as string values only
 ` * added `stubbles\lang\Properties::parseValue()`